Sizing yaml template for SystemLink Enterprise pilots (#217)
* Sizing yaml template for SystemLink Enterprise pilots

Provide a sizing file that specifies resources and replicas for a SystemLink Enterprise pilot deployment.

* Alphabetize service list

Alphabetize service list to make it easier for users to find them

* Improve pilot sizing template instructions

Add detailed comments for the polot-sizing.yaml so the user knows what the file does, how to use it and the intended outcome.

* Remove cluster-internal MongoDB overrides

The default configuration is external MongoDB, even for pilots.  Remove these overrides since they are not needed.

* Comment for production sizings

Add a comment indicating that production deployments should use the defaults provided in the helm charts, or they could be overridden similarly to this file to increase the resources or replicas for specific workloads.

* Use managed MongoDB for pilots

Indicate in the comments that Pilot deployments use managed MongoDB Atlas or MongoDB Enterprise.