Add documentation for FlinkOperator watchNamespaces configuration
We need this because the ni.com public docs are requiring different namespaces
for systemlink and systemlinkadmin. So following those docs without
configuring the watchNamespaces ends up with failures of the
FlinkDeployment component.

All our SL deployments use the same namespaces so this is why
we've seen this so late and it was never an issue.

Signed-off-by: Cristian Hotea <cristian.hotea@emerson.com>